+++ src/nationalarchives/components/checkboxes/checkboxes.stories.js
@@ -21,6 +21,10 @@ export default {
       items: { control: "object" },
       small: { control: "boolean" },
       inline: { control: "boolean" },
+      formItemClasses: { control: "text" },
+      formItemAttributes: { control: "object" },
+      fieldsetClasses: { control: "text" },
+      fieldsetAttributes: { control: "object" },
       classes: { control: "text" },
       attributes: { control: "object" },
     }).map(([key, value]) => [